+++ TravelApp/TravelApp/App.xaml
@@ -2,7 +2,7 @@
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:local="clr-namespace:TravelApp"
-             StartupUri="/MVVM/View/Window1.xaml">
+             StartupUri="/MVVM/View/FormaAranzman.xaml">
     <Application.Resources>
 
         <SolidColorBrush x:Key="TextBox.MouseOver.Border" Color="#3ea8c2"/>
